Filetree created & organized, API returning response
