+++ src/container/App/App.test.js
@@ -1,6 +1,6 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
+import React from './node_modules/react';
+import ReactDOM from './node_modules/react-dom';
+import App from '.';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
